# Accelerated worksheet parsing