--- a/include/xlnt/utils/numeric.hpp
+++ b/include/xlnt/utils/numeric.hpp
@@ -29,6 +29,7 @@
 #include <sstream>
 #include <type_traits>
 #include <cassert>
+#include <algorithm>
 
 namespace xlnt {
 namespace detail {