--- a/source/detail/header_footer/header_footer_code.cpp
+++ b/source/detail/header_footer/header_footer_code.cpp
@@ -22,7 +22,7 @@
 // @author: see AUTHORS file
 
 #include <detail/header_footer/header_footer_code.hpp>
-#include <detail/numeric_utils.hpp>
+//#include <detail/numeric_utils.hpp>
 
 namespace xlnt {
 namespace detail {