@tfussell 2019-11-16
This is great, thanks! Will try to get this merged in soon.