@Crzyrndm 2019-11-17
Have cleaned up and commented the code, much less experimental/hacky than the original PR. There's enough low hanging fruit here that I'm pretty sure load times could be halved again (e.g. by using a background thread to parse the XML and feeding the output into xlnt::worksheet in the main thread I'd estimate another +30-50% throughput is available. Re-doing the rest of the XML parsing would assist with string heavy worksheets. etc.)

It's working at the level I need it though (data processing is now outweighing data loading...), so I'll be stopping here for now (pending review comments). Opened https://github.com/tfussell/xlnt/issues/422 for the number_converter speed issues