@Crzyrndm 2019-11-18
Implemented copy & replace of '.' when in a comma locale as discussed in #422, fixed a number of locations where xlsx_consumer was relying on attribute<double> (which is just stringstream again so had locale AND speed issues)

xlsx_producer is definitely NOT locale aware, since it uses libstudxml for quite a few conversions, but that's a battle for another day