@Crzyrndm 2019-11-19
@tfussell 
somewhat related side note, this library currently has very few external dependencies (just libstudxml?). Doing the work for this PR had me looking at other libs in both supporting and performance roles (e.g. testing / benchmarking ([1](https://github.com/google/googletest), [2](https://github.com/catchorg/Catch2)), alternative non-std containers (particularly hashmaps which this lib uses heavily)

Do you have a strong opinion / guidelines around adding dependencies (my first recommendation after this PR would be google test / benchmark. Both would assist greatly if I was to continue making improvements to xlsx loading)