@tfussell 2019-12-19
@Crzyrndm  Sorry it took so long to get this merged in. As far as dependencies, I'm totally open to that. What I do want to avoid with this library is depending on system libraries (like requiring Ubuntu users to do `apt-get install libxml`). I guess this means vendoring the libraries like I've done with libstdxml.